fix(review): 번역 저장 갱신과 접근성 보완
